Add/Nt/Installer: Loader warning fix and makefile adjustments for arm64. jiraref:VBP-1253
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106713 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/Installer/Loader/VBoxWindowsAdditions.cpp
@@ -968,7 +968,7 @@ int main()
 
     SetLastError(NO_ERROR);
     WCHAR wszExePath[MAX_PATH] = { 0 };
-    DWORD cwcExePath = GetModuleFileNameW(NULL, wszExePath, sizeof(wszExePath));
+    size_t cwcExePath = GetModuleFileNameW(NULL, wszExePath, sizeof(wszExePath));
     if (cwcExePath == 0 || cwcExePath >= sizeof(wszExePath))
         return ErrorMsgRcLastErrSUR(13, "GetModuleFileNameW failed: ", cwcExePath);
